// packages/questions/src/generated/types/publicQuestionBlueprintDocument.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Lemma Questions API
 * OpenAPI spec version: 0.1.0
 */
import type { PublicQuestionBlueprintBlock } from "./publicQuestionBlueprintBlock.js";
import type { PublicQuestionBlueprintDocumentSchemaVersion } from "./publicQuestionBlueprintDocumentSchemaVersion.js";
import type { QuestionResponseField } from "./questionResponseField.js";

export interface PublicQuestionBlueprintDocument {
  blocks: PublicQuestionBlueprintBlock[];
  responseFields: QuestionResponseField[];
  schemaVersion: PublicQuestionBlueprintDocumentSchemaVersion;
}