// packages/questions/src/generated/types/questionBody.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Lemma Questions API
 * OpenAPI spec version: 0.1.0
 */
import type { QuestionBlock } from "./questionBlock.js";
import type { QuestionBodySchemaVersion } from "./questionBodySchemaVersion.js";
import type { QuestionResponseField } from "./questionResponseField.js";

export interface QuestionBody {
  blocks: QuestionBlock[];
  responseFields: QuestionResponseField[];
  schemaVersion: QuestionBodySchemaVersion;
}